requirements/development.txt
black>=19.10b0
flake8>=3.7.8
Flask-Shell-IPython>=0.4.1
Flask-WebTest>=0.0.9
isort>=4.3.21, <5
pycodestyle>=2.5.0
pylint>=2.4.2
pytest>=5.2.0
pytest-cov>=2.7.1
